DB connection User-Recipe-Ingredient

+++ backend/src/main/java/com/foodies/backend/security/user/UserRepository.java
@@ -1,9 +1,12 @@
 package com.foodies.backend.security.user;
 
+import com.foodies.backend.recipeLogic.dbConnection.Recipe;
 import org.springframework.data.jpa.repository.JpaRepository;
 
+import java.util.List;
 import java.util.Optional;
 
-public interface UserRepository extends JpaRepository<User, Integer> {
+public interface UserRepository extends JpaRepository<User, Long> {
     Optional<User> findByUsername(String username);
+
 }
